Rename the test support package
